dropzone executes multiple upload calls despite that we set uploadMultiple, we thus create the session key once before the upload and use it for all uploads till the form is saved

+++ attributes/multi_file/controller.php
@@ -158,7 +158,9 @@ class Controller extends AttributeTypeController
             $removeFiles = preg_split('[,]', $data['removeFiles'], -1, PREG_SPLIT_NO_EMPTY);
             foreach ($removeFiles as $fID) {
                 $file = File::getByID($fID);
-                $file->delete();
+                if (!is_null($file)) {
+                    $file->delete();
+                }
             }
         }
